Refactor code into `influxio.adapter` vs. `influxio.model`
